fixup! fixup! fixup! TW-3070: implement web overlay for video download and enhance file extension handling
