+++ lib/widgets/video_player.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:typed_data';
 
 import 'package:fluffychat/utils/platform_infos.dart';
-import 'package:fluffychat/widgets/video_viewer_web_overlay.dart';
+import 'package:fluffychat/widgets/media_viewer_web_overlay.dart';
 import 'package:flutter/material.dart';
 import 'package:matrix/matrix.dart';
 import 'package:media_kit/media_kit.dart';
@@ -35,7 +35,7 @@ class VideoPlayer extends StatefulWidget {
 class _VideoPlayerState extends State<VideoPlayer> {
   late Player player;
   late VideoController videoController;
-  final _webOverlay = VideoViewerWebOverlay();
+  final _webOverlay = MediaViewerWebOverlay.video();
 
   /// Creates a [Player] with explicit log level so decoder selection warnings
   /// (e.g. silent HW → SW fallback on Android) are captured in logs.
